refactor(parser-jvm-core): move annotation logic to an AnnotatedAbstractModel
